Prompt_Chaining/summarizing_research_papers/pyproject.toml
[project]
name = "summarizing-research-papers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Humza134", email = "mhumza13420@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "langchain-google-genai>=2.1.0",
    "langgraph>=0.3.10",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
]

[project.scripts]
invoke = "summarizing_research_papers:main"
stream = "summarizing_research_papers:stream"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
